+++ affected-paths/core/src/main/kotlin/com/squareup/affected/paths/core/CoreAnalyzer.kt
@@ -79,10 +79,21 @@ public class CoreAnalyzer @JvmOverloads constructor(private val coreOptions: Cor
 
           val projectConnection = with(affectedPathsApplication.koin.get<GradleConnector>()) {
             forProjectDirectory(rootDir.toFile())
-            if (coreOptions.gradleInstallationPath != null) {
-              useInstallation(coreOptions.gradleInstallationPath.toFile())
-            } else {
-              useBuildDistribution()
+            when {
+              coreOptions.gradleDistributionPath != null -> {
+                useDistribution(coreOptions.gradleDistributionPath.toUri())
+              }
+              coreOptions.gradleInstallationPath != null -> {
+                useInstallation(coreOptions.gradleInstallationPath.toFile())
+              }
+              coreOptions.gradleVersion != null -> {
+                useGradleVersion(coreOptions.gradleVersion)
+              }
+              else -> useBuildDistribution()
+            }
+
+            if (coreOptions.gradleUserHome != null) {
+              useGradleUserHomeDir(coreOptions.gradleUserHome.toFile())
             }
             return@with connect()
           }
@@ -96,6 +107,9 @@ public class CoreAnalyzer @JvmOverloads constructor(private val coreOptions: Cor
             )
           )
           actionExecutor.withCancellationToken(cancellationTokenSource.token())
+          if (coreOptions.useBuildScan) {
+            actionExecutor.forTasks(emptyList())
+          }
           actionExecutor.addArguments(coreOptions.gradleArgs)
           actionExecutor.addJvmArguments(coreOptions.jvmArgs)