+++ affected-paths/core/src/main/kotlin/com/squareup/affected/paths/core/CoreOptions.kt
@@ -70,8 +70,33 @@ public data class CoreOptions @JvmOverloads constructor(
   /** Include any "includeBuild" builds from the current build */
   val useIncludeBuild: Boolean = true,
 
-  /** Pass in a custom Gradle installation, instead of using the build distribution */
+  /** Gradle distribution file to use */
+  val gradleDistributionPath: Path? = null,
+
+  /**
+   * Pass in a custom Gradle installation, instead of using the build distribution
+   *
+   * **NOTE**: This will override `gradleDistributionPath` if used.
+   */
   val gradleInstallationPath: Path? = null,
+
+  /**
+   * Gradle version to use for the current build
+   *
+   * **NOTE**: This will override `gradleInstallationPath` and `gradleDistributionPath` if used.
+   */
+  val gradleVersion: String? = null,
+
+  /** Gradle user home directory to use */
+  val gradleUserHome: Path? = null,
+
+  /**
+   * Add the build scan flag to the tooling.
+   *
+   * **Note**: This will cause the default tasks of a build to run.
+   */
+
+  val useBuildScan: Boolean = false,
 ) {
 
   init {
@@ -108,7 +133,7 @@ public data class CoreOptions @JvmOverloads constructor(
                     mavenCentral()
                   }
                   dependencies {
-                    classpath "com.squareup.affected.paths:tooling-support:0.1.3"
+                    classpath "com.squareup.affected.paths:tooling-support:0.1.4"
                   }
                 }
               }
@@ -122,6 +147,9 @@ public data class CoreOptions @JvmOverloads constructor(
           deleteOnExit()
         }.absolutePath
       )
+    }
+    if (useBuildScan) {
+      add("--scan")
     }
   }
 }